Arel::Nodes::CurrentOfExpression
