+++ lib/arel/extensions/nodes.rb
@@ -4,17 +4,6 @@
 
 module Arel
   module Nodes
-    # https://www.postgresql.org/docs/10/sql-update.html
-    class CurrentOfExpression < Arel::Nodes::Node
-      attr_accessor :cursor_name
-
-      def initialize(cursor_name)
-        super()
-
-        @cursor_name = cursor_name
-      end
-    end
-
     # https://www.postgresql.org/docs/9.5/sql-insert.html
     Arel::Nodes::DeleteStatement.class_eval do
       attr_accessor :using
